Rename `ai-model-provider` to `ai-core`
