+++ examples/browser-only/tsconfig.json
@@ -15,7 +15,7 @@
       "path": "../../packages/ai-chat-ui"
     },
     {
-      "path": "../../packages/ai-model-provider"
+      "path": "../../packages/ai-core"
     },
     {
       "path": "../../packages/ai-openai"
