+++ examples/electron/tsconfig.json
@@ -18,7 +18,7 @@
       "path": "../../packages/ai-chat-ui"
     },
     {
-      "path": "../../packages/ai-model-provider"
+      "path": "../../packages/ai-core"
     },
     {
       "path": "../../packages/ai-openai"
