+++ packages/ai-core/.eslintrc.js
