+++ packages/ai-core/README.md
