+++ packages/ai-core/src/browser/ai-core-frontend-module.ts
