+++ packages/ai-core/tsconfig.json
